+++ app/src/main/AndroidManifest.xml
@@ -54,10 +54,10 @@
             android:name=".SearchMapsActivity"
             android:label="@string/title_activity_search_maps" />
         <activity android:name=".SearchActivity" />
-        <activity android:name=".RiderViewRequestActivity"/>
         <activity android:name=".HelpActivity" />
         <activity android:name=".MakeRequestActivity"
             android:windowSoftInputMode="adjustPan"/>
+        <activity android:name=".RiderRequestActivity" />
     </application>
 
 </manifest>
